requirements.txt
matplotlib==3.7.2
numpy==1.25.2
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchsummary==1.5.1
torchvision==0.13.0+cu113